refactor to support other LM tasks
